Add escrow-example-1 Solana escrow
* Objective

This commit objective is to add my first escrow example.

* Why

I'm commited to learning Solana & Rust, and this is the first approach to it.

* How

I'm following the following blogpost:

https://paulx.dev/blog/2021/01/14/programming-on-solana-an-introduction/

And the changes introduced on this commit are a reflection of that
blogpost.
